+++ NEWS.md
@@ -1,6 +1,8 @@
 # cccesMRPprep 0.1.13
 
 * Add 2022, 2024, 2006, and 2012 daily kos cd data
+* Fix minor error where pct values in cd_info from 2006-2012, 2022-2024 were not using two-party voteshare
+* cd_info now distinguishes between total votes including third party (`presvotes_total`) and D+R votes (`presvotes_DR`)
 
 # cccesMRPprep 0.1.12